# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movieflix API",
    "description": "API para gerenciamento de filmes do Movieflix",
    "version": "1.0.0"
  },
  "paths": {
    "/movies": {
      "get": {
        "tags": ["Movie"],
        "summary": "Obter a lista dos filmes",
        "responses": {
          "200": {
            "description": "Retorna a lista de filmes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "put": {
        "tags": ["Movie"],
        "summary": "Atualizar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/movieImput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso"
          },
          "404": {
            "description": "Filme nao encontrado"
          },
          "500": {
            "description": "Falha ao atualizar o filme"
          }
        }
      },
      "delete": {
        "tags": ["Movie"],
        "summary": "Deletar um filme",
        "parameters": [
            
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme a ser removido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filme removido com sucesso"
          },
          "404": {
            "description": "Filme nao encontrado"
          },
          "500": {
            "description": "Não foi possivel remove o filme"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do filme",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "Titulo do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "Genero do filme"
          },
          "genres": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "Id do gênero"
              },
              "name": {
                "type": "string",
                "description": "Nome do gênero"
              }
            }
          },
          "language_id": {
            "type": "integer",
            "description": "Linguagem do filme"
          },
          "languages": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "Id da linguagem"
              },
              "name": {
                "type": "string",
                "description": "Nome do linguagem"
              }
            }
          },
          "oscar_count": {
            "type": "integer",
            "description": "Quantidade de Oscars do filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lancamento do filme"
          }
        }
      },
      "movieImput": {
        "type": "object",
        "required": [
          "title",
          "genre_id",
          "language_id",
          "oscar_count",
          "release_date"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Titulo do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "Genero do filme"
          },
          "language_id": {
            "type": "integer",
            "description": "Linguagem do filme"
          },
          "oscar_count": {
            "type": "integer",
            "description": "Quantidade de Oscars do filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lancamento do filme"
          }
        }
      }
    }
  }
}